# stocktool-kiosk/openapi-cloud-api.yaml
openapi: 3.0.3
info:
  title: StockTool Cloud API
  description: |
    Inventory, tools, projects, and barcode management, plus the
    installation/sync/config/update surface used by the StockTool Kiosk
    (v2) desktop client.

    ## Authentication
    Two separate credential types are used depending on the caller:

    - **User JWT** (`Authorization: Bearer <jwt>`) — obtained from
      `POST /api/auth/login` or `POST /api/auth/kiosk-login`. Used by the
      admin frontend and any endpoint acting on behalf of a specific
      person. Admin-only endpoints additionally require the JWT's user to
      have `role: admin`.
    - **Installation token** (`Authorization: Bearer <token>`) — obtained
      once from `POST /api/installations/register`. Used by the kiosk-v2
      desktop client for `/api/config`, `/api/updates/latest`, and
      `/api/sync/*`. Represents a trusted *device* relaying many local
      users' actions, not one person's session — a different credential
      type from the user JWT on purpose.
  version: "1.0.0"
servers:
  - url: https://api-stocktool.opslabsystems.cloud
    description: Production
...
security:
  - userJwt: []
paths:
...
  /api/users/:
    get:
      tags: [Users]
      summary: List users (admin only)
      parameters:
        - { name: q, in: query, schema: { type: string }, description: "Search by username/email" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/User" } }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [Users]
      summary: Create a user (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username]
              properties:
                username: { type: string }
                email: { type: string, description: "Required unless badge_only is true" }
                password: { type: string, minLength: 6, description: "Required unless badge_only is true" }
                role: { type: string, enum: [admin, stock_user] }
                force_password_change: { type: boolean }
                badge_only:
                  type: boolean
                  description: >
                    Badge-only accounts have no password/email and can only log
                    in via kiosk badge scan; role is forced to stock_user.
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/users/{user_id}:
    get:
      tags: [Users]
      summary: Get a user (admin only)
      parameters: [{ $ref: "#/components/parameters/UserId" }]
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/User" } } } }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [Users]
      summary: Update a user (admin only)
      parameters: [{ $ref: "#/components/parameters/UserId" }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string }
                role: { type: string, enum: [admin, stock_user] }
                is_active: { type: boolean }
                force_password_change: { type: boolean }
                new_password: { type: string, minLength: 6 }
                badge_only: { type: boolean, description: "Set true to convert to badge-only (clears password); set false + new_password to convert back" }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/User" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "409": { description: "Would remove the last active admin" }
    delete:
      tags: [Users]
      summary: Delete a user (admin only; disabled instead if they have history)
      parameters: [{ $ref: "#/components/parameters/UserId" }]
      responses:
        "200": { description: Deleted or disabled }
        "400": { description: "Cannot delete your own account" }
        "409": { description: "Cannot delete the last admin, or user has tools checked out" }

  # ── Items ───────────────────────────────────────────────────────────
  /api/items/:
    get:
      tags: [Items]
      summary: List items
      parameters:
        - { name: q, in: query, schema: { type: string } }
        - { name: category, in: query, schema: { type: string } }
        - { name: stock, in: query, schema: { type: string, enum: [low, out] } }
      responses:
        "200": { description: OK, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/Item" } } } } }
    post:
      tags: [Items]
      summary: Create an item (admin only)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                sku: { type: string }
                description: { type: string }
                category: { type: string }
                location: { type: string }
                quantity: { type: integer, default: 0 }
                unit: { type: string }
                low_stock_threshold: { type: integer, default: 5 }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/Item" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
  /api/items/{item_id}:
    get:
      tags: [Items]
      summary: Get an item
      parameters: [{ $ref: "#/components/parameters/ItemId" }]
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Item" } } } }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [Items]
      summary: Update an item (admin only)
      parameters: [{ $ref: "#/components/parameters/ItemId" }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                sku: { type: string }
                description: { type: string }
                category: { type: string }
                location: { type: string }
                unit: { type: string }
                low_stock_threshold: { type: integer }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Item" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
    delete:
      tags: [Items]
      summary: Soft-delete an item (admin only)
      parameters: [{ $ref: "#/components/parameters/ItemId" }]
      responses:
        "200": { description: Deleted }
  /api/items/{item_id}/adjust:
    post:
      tags: [Items]
      summary: Adjust stock by a signed delta
      parameters: [{ $ref: "#/components/parameters/ItemId" }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [delta]
              properties:
                delta: { type: integer, description: "Positive to add, negative to remove; cannot be 0" }
                notes: { type: string }
                device: { type: string }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Item" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
...
components:
  securitySchemes:
    userJwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
    installToken:
      type: http
      scheme: bearer
      description: Per-installation token from POST /api/installations/register
  parameters:
    UserId: { name: user_id, in: path, required: true, schema: { type: integer } }
    ItemId: { name: item_id, in: path, required: true, schema: { type: integer } }
...
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema: { type: object, properties: { error: { type: string } } }
...
    Forbidden:
      description: Authenticated but not permitted (e.g. non-admin calling an admin-only endpoint)
      content:
        application/json:
          schema: { type: object, properties: { error: { type: string } } }
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { type: object, properties: { error: { type: string } } }
  schemas:
    User:
      type: object
      properties:
        id: { type: integer }
        username: { type: string }
        email: { type: string, nullable: true }
        role: { type: string, enum: [admin, stock_user] }
        is_admin: { type: boolean }
        is_active: { type: boolean }
        force_password_change: { type: boolean }
        has_password: { type: boolean }
        badge_code: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        last_login: { type: string, format: date-time, nullable: true }
    Item:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        sku: { type: string, nullable: true }
        description: { type: string, nullable: true }
        category: { type: string, nullable: true }
        location: { type: string, nullable: true }
        quantity: { type: integer }
        unit: { type: string, nullable: true }
        low_stock_threshold: { type: integer }
        stock_status: { type: string, enum: [in_stock, low_stock, out_of_stock] }
        is_low_stock: { type: boolean }
        is_active: { type: boolean }
        barcode_code: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...